requirements.txt
click
pyyaml
python-grpc